requirements.txt
yfinance
pandas
boto3
pyarrow
python-dotenv